ml_functions/predictive_models/Random_Forest/requirements.txt
google-cloud-storage
google-cloud-aiplatform
pandas
functions-framework
scikit-learn
google-cloud-bigquery
pyarrow
